curious-ai-backend/requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-tavily
fastapi
uvicorn
python-dotenv
